move pois controller to api namespace
