+++ app/views/api/v0/pois/index.json.jbuilder
