+++ app/views/api/v0/pois/show.json.jbuilder
